- Remove `SettingsRouter`. - Move that logic into HandleLemmyLinks.
