+++ Mlem/Extensions/View - Handle Lemmy Links.swift
@@ -10,6 +10,8 @@ import Foundation
 import SwiftUI
 
 struct HandleLemmyLinksDisplay: ViewModifier {
+    @Environment(\.navigationPath) private var navigationPath
+    @EnvironmentObject private var layoutWidgetTracker: LayoutWidgetTracker
     @EnvironmentObject var appState: AppState
     @EnvironmentObject var filtersTracker: FiltersTracker
     
@@ -18,6 +20,7 @@ struct HandleLemmyLinksDisplay: ViewModifier {
     @AppStorage("upvoteOnSave") var upvoteOnSave = false
 
     // swiftlint:disable function_body_length
+    // swiftlint:disable:next cyclomatic_complexity
     func body(content: Content) -> some View {
         content
             .navigationDestination(for: NavigationRoute.self) { route in
@@ -72,8 +75,101 @@ struct HandleLemmyLinksDisplay: ViewModifier {
                         .environmentObject(appState)
                 }
             }
+            .navigationDestination(for: SettingsRoute.self) { route in
+                switch route {
+                case .accountsPage:
+                    AccountsPage()
+                case .general:
+                    GeneralSettingsView()
+                case .accessibility:
+                    AccessibilitySettingsView()
+                case .appearance:
+                    AppearanceSettingsView()
+                case .contentFilters:
+                    FiltersSettingsView()
+                case .about:
+                    AboutView(navigationPath: navigationPath)
+                case .advanced:
+                    AdvancedSettingsView()
+                case .aboutPage(let path):
+                    aboutPageDestination(for: path)
+                case .appearancePage(let path):
+                    appearancePageDestination(for: path)
+                case .commentPage(let path):
+                    commentPageDestination(for: path)
+                case .postPage(let path):
+                    postPageDestination(for: path)
+                case .licensesPage(let path):
+                    licensesPageDestination(for: path)
+                }
+            }
     }
     // swiftlint:enable function_body_length
+    
+    @ViewBuilder
+    private func aboutPageDestination(for path: AboutSettingsRoute) -> some View {
+        switch path {
+        case .contributors:
+            ContributorsView()
+        case let .eula(doc):
+            DocumentView(text: doc.body)
+        case let .privacyPolicy(doc):
+            DocumentView(text: doc.body)
+        case .licenses:
+            LicensesView()
+        }
+    }
+    
+    @ViewBuilder
+    private func appearancePageDestination(for path: AppearanceSettingsRoute) -> some View {
+        switch path {
+        case .theme:
+            ThemeSettingsView()
+        case .appIcon:
+            IconSettingsView()
+        case .posts:
+            PostSettingsView()
+        case .comments:
+            CommentSettingsView()
+        case .communities:
+            CommunitySettingsView()
+        case .users:
+            UserSettingsView()
+        case .tabBar:
+            TabBarSettingsView()
+        }
+    }
+    
+    @ViewBuilder
+    private func commentPageDestination(for path: CommentSettingsRoute) -> some View {
+        switch path {
+        case .layoutWidget:
+            LayoutWidgetEditView(widgets: layoutWidgetTracker.groups.comment, onSave: { widgets in
+                layoutWidgetTracker.groups.comment = widgets
+                layoutWidgetTracker.saveLayoutWidgets()
+            })
+        }
+    }
+    
+    @ViewBuilder
+    private func postPageDestination(for path: PostSettingsRoute) -> some View {
+        switch path {
+        case .customizeWidgets:
+            /// We really should be passing in the layout widget through the route enum value, but that would involve making layout widget tracker hashable and codable.
+            LayoutWidgetEditView(widgets: layoutWidgetTracker.groups.post, onSave: { widgets in
+                layoutWidgetTracker.groups.post = widgets
+                layoutWidgetTracker.saveLayoutWidgets()
+            })
+        }
+    }
+    
+    @ViewBuilder
+    private func licensesPageDestination(for path: LicensesSettingsRoute) -> some View {
+        switch path {
+        case let .licenseDocument(doc):
+            DocumentView(text: doc.body)
+        }
+    }
 }
 
 struct HandleLemmyLinkResolution<Path: AnyNavigationPath>: ViewModifier {
